+++ src/Vector3.cc
@@ -342,9 +342,11 @@ bool IGN_VECTOR3::operator!=(const IGN_VECTOR3 &_pt) const
 //////////////////////////////////////////////////
 bool IGN_VECTOR3::IsFinite() const
 {
-  return std::isfinite(this->data[0]) &&
-         std::isfinite(this->data[1]) &&
-         std::isfinite(this->data[2]);
+  // std::isfinite works with floating point values, need to explicit cast to
+  // avoid ambiguity in vc++.
+  return std::isfinite(static_cast<double>(this->data[0])) &&
+         std::isfinite(static_cast<double>(this->data[1])) &&
+         std::isfinite(static_cast<double>(this->data[2]));
 }
 
 //////////////////////////////////////////////////
