+++ src/Vector4.cc
@@ -258,8 +258,12 @@ bool IGN_VECTOR4::operator!=(const IGN_VECTOR4 &pt) const
 //////////////////////////////////////////////////
 bool IGN_VECTOR4::IsFinite() const
 {
-  return std::isfinite(this->data[0]) && std::isfinite(this->data[1]) &&
-    std::isfinite(this->data[2]) && std::isfinite(this->data[3]);
+  // std::isfinite works with floating point values, need to explicit cast to
+  // avoid ambiguity in vc++.
+  return std::isfinite(static_cast<double>(this->data[0])) && 
+         std::isfinite(static_cast<double>(this->data[1])) &&
+         std::isfinite(static_cast<double>(this->data[2])) && 
+         std::isfinite(static_cast<double>(this->data[3]));
 }
 
 //////////////////////////////////////////////////
